+++ R/tw_caching.R
@@ -63,19 +63,118 @@ tw_set_cache_folder <- function(path = NULL) {
 tw_get_cache_folder <- tw_set_cache_folder
 
 
+
+#' Set database connection settings for the session
+#'
+#' @param db_settings A list of database connection settings (see example)
+#' @param driver A database driver. Common database drivers include `MySQL`, `PostgreSQL`, and `MariaDB`. See `unique(odbc::odbcListDrivers()[[1]])` for a list of locally available drivers.
+#' @param host Host address, e.g. "localhost".
+#' @param port Port to use to connect to the database.
+#' @param database Database name.
+#' @param user Database user name.
+#' @param pwd Password for the database user.
+#'
+#' @return A list with all given parameters (invisibly).
+#' @export
+#'
+#' @examples
+#' \donttest{
+#' if (interactive()) {
+#'
+#' # Settings can be provided either as a list
+#' db_settings <- list(
+#'  driver = "MySQL",
+#'  host = "localhost",
+#'  port = 3306,
+#'  database = "tidywikidatar",
+#'  user = "secret_username",
+#'  pwd = "secret_password"
+#'  )
+#'
+#' tw_set_cache_db(db_settings)
+#'
+#' # or as parameters
+#'
+#' tw_set_cache_db(
+#'  driver = "MySQL",
+#'  host = "localhost",
+#'  port = 3306,
+#'  database = "tidywikidatar",
+#'  user = "secret_username",
+#'  pwd = "secret_password"
+#'  )
+#'
+#' }
+#' }
+tw_set_cache_db <- function(db_settings = NULL,
+                            driver = NULL,
+                            host = NULL,
+                            port,
+                            database,
+                            user,
+                            pwd) {
+  if (is.null(db_settings)==TRUE) {
+    if (is.null(driver)==FALSE) Sys.setenv(tw_db_driver = driver)
+    if (is.null(host)==FALSE) Sys.setenv(tw_db_host = host)
+    if (is.null(port)==FALSE) Sys.setenv(tw_db_port = port)
+    if (is.null(database)==FALSE) Sys.setenv(tw_db_database = database)
+    if (is.null(user)==FALSE) Sys.setenv(tw_db_user = user)
+    if (is.null(pwd)==FALSE) Sys.setenv(tw_db_pwd = pwd)
+    return(invisible(
+      list(driver = driver,
+           host = host,
+           port = port,
+           database = database,
+           user = user,
+           pwd = pwd)))
+  } else {
+    Sys.setenv(tw_db_driver = db_settings$driver)
+    Sys.setenv(tw_db_host = db_settings$host)
+    Sys.setenv(tw_db_port = db_settings$port)
+    Sys.setenv(tw_db_database = db_settings$database)
+    Sys.setenv(tw_db_user = db_settings$user)
+    Sys.setenv(tw_db_pwd = db_settings$pwd)
+    return(invisible(db_settings))
+  }
+}
+
+#' Get database connection settings from the environment
+#'
+#' Typically set with `tw_set_cache_db()`
+#'
+#' @return A list with all database parameters as stored in environment variables.
+#' @export
+#'
+#' @examples
+#'
+#' tw_get_cache_db()
+#'
+tw_get_cache_db <- function() {
+  list(driver = Sys.getenv("tw_db_driver"),
+       host = Sys.getenv("tw_db_host"),
+       port = Sys.getenv("tw_db_port"),
+       database = Sys.getenv("tw_db_database"),
+       user = Sys.getenv("tw_db_user"),
+       pwd = Sys.getenv("tw_db_pwd")
+  )
+}
+
+
 #' Enable caching for the current session
+#'
+#' @param SQLite Logical, defaults to TRUE. Set to FALSE to use custom database options. See `tw_set_cache_db()` for details.
 #'
 #' @return Nothing, used for its side effects.
 #' @export
-
 #' @examples
 #' \donttest{
 #' if (interactive()) {
 #'   tw_enable_cache()
 #' }
 #' }
-tw_enable_cache <- function() {
+tw_enable_cache <- function(SQLite = TRUE) {
   Sys.setenv(tw_cache = TRUE)
+  Sys.setenv(tw_cache_SQLite = SQLite)
 }