+++ NEWS.md
@@ -48,5 +48,6 @@
 * introduce `tw_get_p()` as an alias of `tw_get_property_same_length()` for brevity
 * introduce new parameters in `tw_get_p()` to deal with common pattern when only "preferred" or most recent property should be returned, rather than whatever Wikidata has first in the list, and add new section in the readme
 * add example datasets for illustrative purposes and forthcoming additional vignettes and examples, `tw_qid_meps` and `tw_qid_airports`
-* drop legacy `include_id_and_p` parameter `tw_get_qualifiers()` 
+* drop legacy `include_id_and_p` parameter `tw_get_qualifiers()`
+* add support for setting database connection parameters with `tw_set_cache_db()` for easier use of alternatives to SQLite
 * bug fix: fix error when Wikidata item has no label in any language